+++ CHANGELOG.md
@@ -26,3 +26,5 @@
 - [feat:support zuul circuit breaker fallback response.](https://github.com/Tencent/spring-cloud-tencent/pull/1522)
 - [feat:support auth.](https://github.com/Tencent/spring-cloud-tencent/pull/1541)
 - [feat:support smooth upgrade from tsf. ](https://github.com/Tencent/spring-cloud-tencent/pull/1542)
+- [feat:support polaris event.](https://github.com/Tencent/spring-cloud-tencent/pull/1543)
+